Introduce dependency_container
It's similar to a dependency injection container, but also has
elements of a data store/object mapper over a data store. It
contains little no logic but gives some flexibility when it comes to
introducing new collaborator objects in the future. It is also not a
single as is common with dependency injection containers, but must
instead be passed around.